requirements.txt
pandas==1.5.3
requests==2.28.2
boto3==1.26.0
psycopg2-binary>=2.9,<3.0
sqlalchemy>=1.4,<2.0
apache-airflow==2.10.4
apache-airflow-providers-openlineage>=1.8.0
apache-airflow-providers-telegram==4.7.0
apache-airflow-providers-celery>=3.3.0
